ROMS.create_grid — Method

Non-rotated grids

ROMS.gebco_load — Method

x,y,b = gebco_load(bath_name,xr,yr)

Loads GEBCO bathymetry with lon and lat range xr, yr.

ROMS.interp_clim — Method

interpolate climatology

ROMS.prepare_ecmwf — Method

prepare_ecmwf(atmo_fname,Vnames,filename_prefix,domain_name;
                   time_origin = DateTime(1858,11,17)
)

Generate ROMS forcing fields from the ECMWF data file atmo_fname.

Example

datadir = "..."
atmo_fname = joinpath(datadir,"ecmwf_sample_data.nc")
filename_prefix = joinpath(datadir,"liguriansea_")
domain_name = "Ligurian Sea Region"
Vnames = ["sustr","svstr","shflux","swflux","swrad","Uwind","Vwind","lwrad",
    "lwrad_down","latent","sensible","cloud","rain","Pair","Tair","Qair"]
prepare_ecmwf(atmo_fname,Vnames,filename_prefix,domain_name)
)

Based on forcing/d_ecmwf2roms.m

ROMS.reduce_res — Method

 field = reduce_res(field,red);

reduce resolution of field "field" by a factor "red"

ROMS.relative_humidity — Method

https://web.archive.org/web/20200926200733/https://www.weather.gov/media/epz/wxcalc/vaporPressure.pdf

ROMS.smoothgrid — Method

Smooth the topography to get a maximum r factor = rmax Copyright (c) 2002-2006 by Pierrick Penven, GPL

ROMS.stretching — Method

s,C = stretching(Vstretching, theta_s, theta_b, hc, N, kgrid, report)

STRETCHING: Compute ROMS vertical coordinate stretching function

Given vertical terrain-following vertical stretching parameters, this routine computes the vertical stretching function used in ROMS vertical coordinate transformation. Check the following link for details:

https://www.myroms.org/wiki/index.php/Vertical_S-coordinate

On Input:

Vstretching Vertical stretching function: Vstretching = 1, original (Song and Haidvogel, 1994) Vstretching = 2, A. Shchepetkin (UCLA-ROMS, 2005) Vstretching = 3, R. Geyer BBL refinement Vstretching = 4, A. Shchepetkin (UCLA-ROMS, 2010) Vstretching = 5, Quadractic (Souza et al., 2015) thetas S-coordinate surface control parameter (scalar) thetab S-coordinate bottom control parameter (scalar) hc Width (m) of surface or bottom boundary layer in which higher vertical resolution is required during stretching (scalar) N Number of vertical levels (scalar) kgrid Depth grid type logical switch: kgrid = 0, function at vertical RHO-points kgrid = 1, function at vertical W-points report Flag to report detailed information (OPTIONAL): report = false, do not report report = true, report information

On Output:

s S-coordinate independent variable, [-1 <= s <= 0] at vertical RHO- or W-points (vector) C Nondimensional, monotonic, vertical stretching function, C(s), 1D array, [-1 <= C(s) <= 0]

svn Id: stretching.m 996 2020-01-10 04:28:56Z arango $ ========================================================================# Copyright (c) 2002-2020 The ROMS/TOMS Group # Licensed under a MIT/X style license # See License_ROMS.txt Hernan G. Arango # ========================================================================#

ROMS.vapor_pressure — Method

e = vapor_pressure(T)

actual vapor pressure in hPa (millibars) from dewpoint temperature T in degree Celsius using [1]. If T is the air temperature, then e is the saturated vapor pressure.

[1] https://web.archive.org/web/20200926200733/https://www.weather.gov/media/epz/wxcalc/vaporPressure.pdf

ROMS.Grid — Method

grid = ROMS.Grid(grid_fname,opt)

Note grid.z is for an elevation equal to zero.

Example

opt = (
    Tcline = 50,   # m
    theta_s = 5,   # surface refinement
    theta_b = 0.4, # bottom refinement
    nlevels = 32,  # number of vertical levels
    Vtransform  = 2,
    Vstretching = 4,
)
grid = ROMS.Grid(expanduser("~/Models/LS2v/LS2v.nc"),opt)