ROMS

This package allows to setup necessary files for the ROMS ocean model.

Tutorial

This tutorial is for students following the ULiège lecture OCEA0036-1 but might be useful for others as well.

Using for the first time a UNIX/Linux-like environment?

Please pay attention to the difference between upper- and lowercase letters

Presence and absence of white space is also significant

Check out the basic shell commands and this tutorial

Avoid using directories and file names with a space in them, otherwise you need to put the directory in quotes (single or double quotes) or use black-slash (\) in front of the white space. For example, shell command cd My Directory Name does not work, use one of the following instead:

cd "My Directory Name"
cd 'My Directory Name'
cd My\ Directory\ Name

Registration

Please register at:

ROMS (Regional Ocean Modeling System).

CMEMS (Copernicus Marine Environment Monitoring Service)

To generate new forcing fields, register at (optional):

ECMWF (European Centre for Medium-Range Weather Forecasts)

Required software

A preconfigured virtual machine is available here. The account student has the password tritro. In this virtual machine, all software is already pre-installed. The OVA file must be imported in Virtual Box. If you do not use this virtual machine the following software need to be installed:

Julia. Under Linux, you can install Julia with the following shell commands:

cd /opt/
sudo wget https://julialang-s3.julialang.org/bin/linux/x64/1.7/julia-1.7.2-linux-x86_64.tar.gz
sudo tar -xvf julia-1.7.2-linux-x86_64.tar.gz
sudo rm julia-1.7.2-linux-x86_64.tar.gz
sudo ln -s /opt/julia-1.7.2/bin/julia /usr/local/bin/julia

More information is available here.

Under Linux, you need to install also python3-matplotlib for PyPlot. On Debian/Ubuntu, this packages can be installed by this command:

sudo apt install python3-matplotlib

Julia package, PyPlot, NCDatasets, ROMS which can be installed by:

using Pkg
Pkg.add("PyPlot")
Pkg.add("NCDatasets")
Pkg.develop(url="https://github.com/Alexander-Barth/ROMS.jl")

ROMS source. This example uses the version 3.9 of ROMS. We assume that the ROMS source is copied in ~/src/roms:

mkdir ~/src/
cd ~/src/
git clone "https://www.myroms.org/git/src" roms
cd roms
git checkout roms-3.9

For the git command, you will cneed to provide your ROMS username and password.

Other required software typically available from a package manager:

A Fortran 90 compiler (e.g. gfortran)

GNU make

NetCDF (including headers files for development and the tools ncdump, nf-config)

perl

Python and pip

MPI (optional)

git (optional)

Note that all libraries (NetCDF and MPI) must be compiled with the same Fortran compiler.

On Windows, various ways exist to install gfortran, GNU make and other dependencies:

Windows Subsystem for Linux

Installation guide

FAQ (in particular exchanging files)

Cygwin

MINGW

Linux virtual machine using e.g. VirtualBox

...

On MacOS X:

Homebrew https://brew.sh/

...

On Debian/Ubuntu, these packages can be installed by this command:

sudo apt install gfortran make perl netcdf-bin libnetcdff-dev libopenmpi-dev openmpi-bin git python3-pip python3-setuptools unzip

For CMEMS data, you need the python package motuclient (installation instructions).

For example:

# to install a specific version use
python3 -m pip install --user motuclient==1.8.6
# or to install the latest version use:
# python3 -m pip install --user motuclient

I advice you to use version 1.8.6 of motuclient because of this issue. Normally you will see the warning WARNING: The script motuclient is installed in '.../.local/bin' which is not on PATH. Consider adding this directory to PATH. You need to add the following line to the file .bashrc located in your home directory (at the end of this file on a separate line):

export PATH="$HOME/.local/bin:$PATH"

In a terminal execute the following so that this change takes effect:

source ~/.bashrc

For ECMWF data, you need the pacakge ecmwf-api-client-python (optional). Follow the installation instructions (including the ECMWF key). For questions related to ECMWF data access please also consult this document.

Note that the ECMWF key is different from your ECMWF password.

Check your environment

Check julia version:

julia --version

Check the Fortran Compiler

gfortran --version

MPI

mpif90 --version
mpirun --help

NetCDF

nf-config --all

or the old name nc-config:

nc-config --all

If you have nc-config (with Fortran support) but not nf-config, you can run the following:

ln -s /usr/bin/nc-config $HOME/bin/nf-config

Or set later the environemt variable NF_CONFIG to nc-config.

Check the motuclient (it may return vUnknown, but it should not return No module named motuclient)

motuclient --version

These commands should return a basic usage info or the version number if they are correctly installed.

Data

The full GEBCO bathymetry (the file gebco_30sec_1.nc is already included in the virtual machine)

Area

Choose an area:

What interesting processes are present in your studied area? (note: now we all use the Ligurian Sea)

Are there in situ measurements available for your area? Look for temperature and salinity within your areas (for any time frame)

Check with World Ocean Database

CMEMS in situ Thematic Assemble Centre (TAC)

Visualize a couple of downloaded profiles:

Hints: in Julia you can use the package NCDatasets

How would you distribute the vertical resolution in your model represent this profile?

Check that your longitude/latitude/time range model of your nested model is within longitude/latitude/time range of the outer model providing the boundary conditions

Choose the domain such to avoid unnecessary open ocean boundary conditions

Atmospheric forcing fields

For the Ligurian Sea, necessary parameters have already been prepared and are available in the file ROMS-implementation.zip. It containts data download the from

the ECMWF operational archive (Atmosphere/ecmwf_operational_archive_2018-12-01T00:00:00_2020-01-01T00:00:00.nc). This NetCDF file needs to be converted by the julia function ROMS.prepare_ecmwf.

The remaining of this section explained how to download data from the ECMWF operational archive (e.g. for a different domain). These instructions are not needed now.

Adapt the file name, longitude/latitude and time range (start one day earlier, and finish one day later) in forcing_ecmwf.py and execute the script as follows:

python3 forcing_ecmwf.py

List of variables (*: quantities accumulated over the integration period ("step"))

NetCDF Variable name	Description

msl	Mean sea level pressure

u10	10 metre U wind component

v10	10 metre V wind component

t2m	2 metre temperature

d2m	2 metre dewpoint temperature

tcc	Total cloud cover*

sshf	Surface sensible heat flux*

strd	Surface thermal radiation downwards*

ssr	Surface net solar radiation*

str	Surface net thermal radiation*

ewss	Eastward turbulent surface stress*

nsss	Northward turbulent surface stress*

e	Evaporation*

ro	Runoff*

tp	Total precipitation*

sst	Sea surface temperature

par	Photosynthetically active radiation at the surface*

Generate initial, boundary conditions and forcing files

Adapt the example_config.jl file and call it yourdomain_config.jl where you replace yourdomain by the the name of your domain (lowercase and without space). For the Ligurian Sea, use liguriansea_config.jl.

Longitude/latitude bounding box

File paths

Time range

...

For CMEMS boundary conditions:

You may need to adapt service_id, motu_server and mapping (if model is outside the Mediterranean Sea)

Data will be downloaded and saved in NetCDF by "chunks" of 60 days in the folder OGCM under the content of the variable basedir

You need to remove the files in this directory if you rerun the script with a different time range.

Run in Julia

include("yourdomain_config.jl")

Check the resulting files: initial conditions, boundary conditions, interpolated model (clim file) and visualize the these files along some sections.

ROMS compilation

Create a directory (avoid directory names with spaces) for your model configuration

Compile ROMS:

configure ROMS by creating a file yourdomain.h (e.g. liguriansea.h for the Ligurian Sea):

#define UV_ADV                    /* turn ON advection terms */
#define UV_COR                    /* turn ON Coriolis term */
#define DJ_GRADPS                 /* Splines density  Jacobian (Shchepetkin, 2000) */
#define NONLIN_EOS                /* define if using nonlinear equation of state */
#define SALINITY                  /* define if using salinity */
#define SOLVE3D                   /* define if solving 3D primitive equations */
#define MASKING                   /* define if there is land in the domain */
#define TS_U3HADVECTION           /* Third-order upstream horizontal advection of tracers */
#define TS_C4VADVECTION           /* Fourth-order centered vertical advection of tracers */
#define TS_DIF2                   /* use to turn ON or OFF harmonic horizontal mixing  */
#define MIX_S_UV                  /* mixing along constant S-surfaces */
#define UV_VIS2                   /* turn ON Laplacian horizontal mixing */
#define AVERAGES
#define UV_QDRAG
#define MIX_S_TS

#define  MY25_MIXING
#ifdef MY25_MIXING
# define N2S2_HORAVG
# define KANTHA_CLAYSON
#endif

#define ANA_BSFLUX                /* analytical bottom salinity flux */
#define ANA_BTFLUX                /* analytical bottom temperature flux */
#define ANA_SSFLUX

#define BULK_FLUXES               /* turn ON bulk fluxes computation */
#define CLOUDS
#define LONGWAVE
#define SOLAR_SOURCE

Copy the script build_roms.sh to the directory ~/ROMS-implementation-test

cd ~/ROMS-implementation-test
cp ~/src/roms/ROMS/Bin/build_roms.sh build_roms.sh

Copy it to this directory and adapt it. Here is a list of changes that I made highlighted with the diff tool.

 diff -u /home/abarth/src/roms/ROMS/Bin/build_roms.sh build_roms.sh
--- /home/abarth/src/roms/ROMS/Bin/build_roms.sh	2020-10-11 21:19:08.000000000 +0200
+++ build_roms.sh	2020-11-06 11:16:15.484841910 +0100
@@ -102,12 +102,12 @@
 # determine the name of the ".h" header file with the application
 # CPP definitions.

-export   ROMS_APPLICATION=UPWELLING
+export   ROMS_APPLICATION=LigurianSea

 # Set a local environmental variable to define the path to the directories
 # where all this project's files are kept.

-export        MY_ROOT_DIR=${HOME}/ocean/repository
+export        MY_ROOT_DIR=${HOME}
 export     MY_PROJECT_DIR=${PWD}

 # The path to the user's local current ROMS source code.
@@ -120,7 +120,7 @@
 # machine. This script is designed to more easily allow for differing paths
 # to the code and inputs on differing machines.

- export       MY_ROMS_SRC=${MY_ROOT_DIR}/trunk
+ export       MY_ROMS_SRC=${MY_ROOT_DIR}/src/roms

 # Set path of the directory containing makefile configuration (*.mk) files.
 # The user has the option to specify a customized version of these files
@@ -168,13 +168,13 @@

 #export        USE_OpenMP=on            # shared-memory parallelism

- export              FORT=ifort
-#export              FORT=gfortran
+# export              FORT=ifort
+export              FORT=gfortran
 #export              FORT=pgi

 #export         USE_DEBUG=on            # use Fortran debugging flags
  export         USE_LARGE=on            # activate 64-bit compilation
-#export       USE_NETCDF4=on            # compile with NetCDF-4 library
+export       USE_NETCDF4=on            # compile with NetCDF-4 library
 #export   USE_PARALLEL_IO=on            # Parallel I/O with NetCDF-4/HDF5

If you do not have the tool nf-config, you need to add this line export NF_CONFIG=nc-config.

Review your changes with:

diff ~/src/roms/ROMS/Bin/build_roms.sh build_roms.sh

compile ROMS by running:

./build_roms.sh -j 2

where you need to replace /path/to/previous by the appropriate file path.

copy varinfo.dat from ~/src/roms/ROMS/External/varinfo.dat in your current directory:

cp ~/src/roms/ROMS/External/varinfo.dat .

ROMS model domain configuration

copy roms.in from ~/src/roms/User/External/roms.in in the directory ~/ROMS-implementation-test:

cp  ~/src/roms/User/External/roms.in .

check the glossary at the end of this file for the meaning of the keys that we will change

when editing this file, do not use "tabs".

adapt MyAppCPP and change it to LIGURIANSEA

adapt file names VARNAME, GRDNAME, ININAME, BRYNAME, CLMNAME, FRCNAME and NFFILES

also make sure that these variables are set (number of files with boundary conditions and climatology). If they do not exist, they need to be added (near BRYNAME for example).

 NBCFILES == 1
NCLMFILES == 1

change Lm, Mm and N based on the dimensions of your grid (make sure to read the glossary for these variable in roms.in)

read the desciption about "lateral boundary conditions" and adapt boundaries LBC:

use closed (Clo) for boundaries without sea-point

for open boundaries:

free-surface: Chapman implicit (Cha)

2D U/V-momentum: Flather (Fla)

3D U/V-momentum, temperature, salinity: Radiation with nudging (RadNud)

mixing TKE: Radiation (Rad)

set the starting time and time reference

DSTART = ...
TIME_REF =  18581117

where DSTART is here the number of days since 1858-11-17 or November 17, 1858 (see also modified Julia day). For instance the number of days since 2014-01-01 (year-month-day) can be computed by of following commands in Julia:

using Dates
Date(2020,1,1) - Date(1858,11,17)

The inverse operation can be done with:

using Dates
Date(1858,11,17) + Day(58849)

You can use DateTime if you want to specify hour, minutes or seconds.

Adapt the length of a time step DT (in seconds) and number of time steps NTIMES

Initially we choose:

NTIMES -> number of time step corresponding to 1 day (e.g. 24*60*60/DT as DT is in seconds)

NHIS, NAVG-> number of time steps corresponding to 1 hour

NRST -> number of time steps correspond to 1 hour

Nudging towards "climatology"

A flow relexation zone can be implemented in ROMS by using the followings settings:

LtracerCLM == T T  ! enable processing of CLIM data
LnudgeTCLM == T T  ! nudge to CLIM data
TNUDG == 2*10.0d0                    ! days

Make nudging on inflow is stronger than on outflow

OBCFAC == 10.0d0                      ! nondimensional

Set also NUDNAME to the file name create by the julia script.

Run ROMS

Run ROMS without MPI

To run ROMS without MPI, one need to disable MPI in build_roms.sh. The ROMS binary will be called romsS and call be called by:

./romsS < roms.in | tee roms.out

Run ROMS with MPI

How many CPU cores does your machine have? You can use the command top in a shell terminal followed by 1.

In build_roms.sh set USE_MPI=on (which is actually the default value)

Recompile ROMS

Change in roms.in the parameters NtileI and NtileJ. The number of CPU cores should be NtileI * NtileJ.

Run ROMS with, e.g.

mpirun -np 4 ./romsM  roms.in | tee roms.out

where 4 is the number of cores to use.

With the command tee the normal screen output will be place in the file roms.out but still be printed on the screen.

A problem? What does the error message say?

Outputs are in roms_his.nc and roms_avg.nc, plot some variables like sea surface height and sea surface temperature at the beginning and the end of the simulation.

In Julia, force figure 1 and to 2 to have the same color-bar.

figure(); p1 = pcolor(randn(3,3)); colorbar()
figure(); p2 = pcolor(randn(3,3)); colorbar()
p2.set_clim(p1.get_clim())

If everything runs fine,

is the model still stable with a longer time steps (DT) ?

increase the number of time steps (NTIMES)

possibly adapt the frequency at which you save the model results (NHIS, NAVG,NRST)

Interpreting ROMS output

Check minimum and maximum value of the different parameters

 NLM: GET_STATE - Read state initial conditions,             t = 57235 00:00:00
                   (Grid 02, File: roms_nest_his.nc, Rec=0182, Index=1)
                - free-surface
                   (Min = -4.63564634E-01 Max = -3.63838434E-01)

The barotropic, baroclinic and Coriolis Courant number should be smaller than 1

 Minimum barotropic Courant Number =  2.09670689E-02
 Maximum barotropic Courant Number =  5.56799674E-01
 Maximum Coriolis   Courant Number =  1.71574766E-03

Information

energy (kinetic, potential, total) and volume

maximum Courant number

   STEP   Day HH:MM:SS  KINETIC_ENRG   POTEN_ENRG    TOTAL_ENRG    NET_VOLUME  Grid
          C => (i,j,k)       Cu            Cv            Cw         Max Speed

 346200 57235 00:00:00  2.691184E-03  1.043099E+04  1.043099E+04  6.221264E+13  01
          (079,055,30)  9.266512E-02  4.949213E-02  0.000000E+00  1.081862E+00

Validation

Check out satellite data (e.g. sea surface temperature, sea surface height) at:

CMEMS

PODAAC NASA

Make some comparison with satellite and the downloaded in situ observation

Hydrodynamic model troubleshooting

Hydrodynamic model troubleshooting

More information

ROMS Wiki

ROMS Wiki Frequently Asked Questions

K. Hedström. 2016. Technical Manual for a Coupled Sea-Ice/Ocean Circulation Model (Version 4). U.S. Dept. of the Interior, Bureau of Ocean Energy Management, Alaska OCS Region. OCS, Study BOEM 2016-037. 176 pp.

Reference

ROMS.create_grid — Method

ROMS.create_grid(fname,h,f,lon_r,lat_r,mask_r,angle,pm,pn,dndx,dmde)

Create a NetCDF grid file fname using the bathymetry h, Coriolis parameter f and longitude, latitude, mask, angle and strechting factors are rho-points.

Note

This function currently only works for non-rotated grids (angle = 0) and the spherical grids.

ROMS.diffusion2! — Function

 diffusion2!(f,alpha,Niter)

Two-dimensional diffusion of field f. f is the initial condition of field. alpha is a vector or tuple with two elements corresponding to the diffusion coefficient (multiplied by grid spacing) for the two dimensions. Niter is the number of iterations. On output, f is the field after Niter iterations.

ROMS.extract_bc — Method

ROMS.extract_bc(domain,clim_filename,bc_filename; missing_value = 9999)

From the climatology clim_filename extract the boundary conditions at all open boundaries (using the mask in domain) and save the result into the netCDF nc_filename.

ROMS.extract_ic — Method

ROMS.extract_ic(domain,clim_filename,icfile,t0::DateTime;
                time_origin = DateTime(1858,11,17))

From the climatology clim_filename extract a single time instance at the time t0 (or the nearest) and save the result into icfile.

ROMS.gebco_load — Method

x,y,b = gebco_load(bath_name,xr,yr)

Loads GEBCO bathymetry with lon and lat range xr, yr.

ROMS.generate_grid — Method

ROMS.generate_grid(grid_fname,bath_name,xr,yr,red,opt,hmin,rmax)

Generate the netCDF file grid_fname representing the model grid from the bathymetry file bath_name. The domain is bounded by the longitude range xr and the latitude range yr. The resolution of the bathymetry is reduced by the factors red[1] and red[2] in the longitude and latitude directions. opt is tuple with additional parameters describing the vertical axis. hmin and rmax is the minimum depth and rmax is the smoothness parameter of the bathymetry. The $r$ parameter is defined as:

$r = \max\left( \underset{ij}{\max} \frac{|h_{i,j} - h_{i+1,j}|}{h_{i,j} + h_{i+1,j}}, \underset{ij}{\max} \frac{|h_{i,j} - h_{i,j+1}|}{h_{i,j} + h_{i,j+1}} \right)$

The parameter opt contains for example:

# model specific parameters
opt = (
    Tcline = 50,   # m
    theta_s = 5,   # surface refinement
    theta_b = 0.4, # bottom refinement
    nlevels = 32,  # number of vertical levels
    Vtransform  = 2,
    Vstretching = 4,
)

ROMS.interp_clim — Method

ROMS.interp_clim(domain,clim_filename,dataset,timerange;
                 padding = 0.5,
                 missing_value = -9999.)

Interpolate dataset on the the model grid domain and creating the climatology file clim_filename for all dates between timerange[1] and timerange[2].

ROMS.load — Method

v,(x,y,z,t) = ROMS.load(ds::CMEMS,name::Symbol; kwargs...)

Loads a variable from a CMEMS remote resource. name is the NetCDF CF standard name.

ROMS.nudgecoef — Method

nudgecoef(domain,nudge_filename,alpha,Niter,halo,tscale; max_tscale = 5e5)

Generate trace nudging coefficients with a value of 1/tscale at the open boundaries over a halo grid cells. coefficients field smoothed spatially with a diffusion coefficient alpha over Niter iterations. For grid cells where the nudging time scale exceeds max_tscale, nudging is disabled (coefficient is set to zero).

ROMS.prepare_ecmwf — Method

ROMS.prepare_ecmwf(
   atmo_fname,Vnames,filename_prefix,domain_name;
   time_origin = DateTime(1858,11,17),
   reset_accumulation = time -> Time(time) in (Time(0,0),Time(12,0)),
)

Generate ROMS forcing fields from the ECMWF data file atmo_fname. Note that some variables are accumulated. Per default, the accumulation is reset at 00:00 and 12:00 UTC. The accumulation period is determined from the time resolution (usually 3 hours).

Example

datadir = "..."
atmo_fname = joinpath(datadir,"ecmwf_sample_data.nc")
filename_prefix = joinpath(datadir,"liguriansea_")
domain_name = "Ligurian Sea Region"
Vnames = ["sustr","svstr","shflux","swflux","swrad","Uwind","Vwind","lwrad",
    "lwrad_down","latent","sensible","cloud","rain","Pair","Tair","Qair"]
prepare_ecmwf(atmo_fname,Vnames,filename_prefix,domain_name)
)

Based on forcing/d_ecmwf2roms.m (svn revision 1102):

Copyright (c) 2002-2017 The ROMS/TOMS Group      Hernan G. Arango
Licensed under a MIT/X style license             John Wilkin
See License_ROMS.txt

ROMS.reduce_res — Method

 field = reduce_res(field,red);

reduce resolution of field "field" by a factor "red"

ROMS.relative_humidity — Method

rh = relative_humidity(temperature_2m_C,dew_temperature_2m_C)

Compute the relative humidity (between 0 and 100) from temperature at 2 m, and dew_temperature at 2 m) both in degree Celsius)

[1] https://web.archive.org/web/20200926200733/https://www.weather.gov/media/epz/wxcalc/vaporPressure.pdf

ROMS.smoothgrid — Method

hsmooth = smoothgrid(h,hmin,rmax)

Smooth the topography to get a maximum r factor equalt to rmax

The original code is from Pierrick Penven from Roms_tools. Copyright (c) 2002-2006 by Pierrick Penven, GPL

ROMS.stretching — Method

s,C = stretching(Vstretching, theta_s, theta_b, hc, N, kgrid, report)

Compute ROMS vertical coordinate stretching function.

Given vertical terrain-following vertical stretching parameters, this routine computes the vertical stretching function used in ROMS vertical coordinate transformation. Check the following (link)[https://www.myroms.org/wiki/index.php/Vertical_S-coordinate] for details:

On Input:

Vstretching: Vertical stretching function:

                   Vstretching = 1,  original (Song and Haidvogel, 1994)
                   Vstretching = 2,  A. Shchepetkin (UCLA-ROMS, 2005)
                   Vstretching = 3,  R. Geyer BBL refinement
                   Vstretching = 4,  A. Shchepetkin (UCLA-ROMS, 2010)
                   Vstretching = 5,  Quadractic (Souza et al., 2015)

theta_s: S-coordinate surface control parameter (scalar)

theta_b: S-coordinate bottom control parameter (scalar)

hc: Width (m) of surface or bottom boundary layer in which higher vertical resolution is required during stretching (scalar)

N: Number of vertical levels (scalar)

kgrid Depth grid type logical switch: kgrid = 0, function at vertical RHO-points kgrid = 1, function at vertical W-points

report Flag to report detailed information (OPTIONAL): report = false, do not report report = true, report information

On Output:

s: S-coordinate independent variable, [-1 <= s <= 0] at vertical RHO- or W-points (vector)

C: Nondimensional, monotonic, vertical stretching function, C(s), 1D array, [-1 <= C(s) <= 0]

The code is ported from the matlab code stretching.m from Hernan G. Arango. Copyright (c) 2002-2020 The ROMS/TOMS Group, Licensed under a MIT/X style license See LicenseROMS.txt, Hernan G. Arango https://www.myroms.org/main.php?page=LicenseROMS

ROMS.vapor_pressure — Method

e = vapor_pressure(T)

actual vapor pressure in hPa (millibars) from dewpoint temperature T in degree Celsius using using [1]. If T is the air temperature, then e is the saturated vapor pressure over liquid water is given by:

$e(T) = 6.11 \cdot 10 ^ {\left( \frac{7.5 T}{237.7 + T} \right)}$

[1] https://web.archive.org/web/20200926200733/https://www.weather.gov/media/epz/wxcalc/vaporPressure.pdf

ROMS.vapor_pressure_Buck — Method

e = vapor_pressure_Buck(T)

actual vapor pressure in hPa (millibars) from dewpoint temperature T in degree Celsius using using Buck (1996). If T is the air temperature, then e is the saturated vapor pressure over liquid water is given by:

$e(T) = 6.1121 \exp \left(\left( 18.678 - \frac{T} {234.5}\right)\left( \frac{T} {257.14 + T} \right)\right)$

ROMS.CMEMS — Method

ds = ROMS.CMEMS(username,password,service_id,mapping,cachedir;
               motu_server = "http://nrt.cmems-du.eu/motu-web/Motu",
               # Put here the path of the script motuclient
               motu_program = "motuclient",
           )

Returns a structure ds to connect to a CMEMS Motu server using the python tools motuclient (which must be available in your PATH). The mapping parameter contains a dictorary linking the NetCDF CF standard namer to the underlying NetCDF variable names and the product identifers (more information is available in the product user manual). cachedir is a directory where the products are downloaded for caching.

Example

The values of service_id and mapping below are specific to the Mediterranean Sea and must be adapted for other domains.

cmems_username = "Alice"
cmems_password = "rabbit"
outdir = "/tmp"
service_id = "MEDSEA_ANALYSISFORECAST_PHY_006_013-TDS"
mapping = Dict(
    # var  product_id
    :sea_surface_height_above_geoid => ("zos","med-cmcc-ssh-an-fc-d"),
    :sea_water_potential_temperature => ("thetao", "med-cmcc-tem-an-fc-d"),
    :sea_water_salinity => ("so","med-cmcc-sal-an-fc-d"),
    :eastward_sea_water_velocity => ("uo", "med-cmcc-cur-an-fc-d"),
    :northward_sea_water_velocity => ("vo", "med-cmcc-cur-an-fc-d"),
)
dataset = ROMS.CMEMS(cmems_username,cmems_password,service_id,mapping,outdir)

ROMS.Grid — Method

grid = ROMS.Grid(grid_fname,opt)

Loads the ROMS grid from a NetCDF file grid_fname. The grid structure contains the longitude, latitude, depth (z), angle and mask at rho, u, v and w points for a C-grid.

Note

grid.z corresponds to an elevation equal to zero.

Example

opt = (
    Tcline = 50,   # m
    theta_s = 5,   # surface refinement
    theta_b = 0.4, # bottom refinement
    nlevels = 32,  # number of vertical levels
    Vtransform  = 2,
    Vstretching = 4,
)
grid = ROMS.Grid(expanduser("~/Models/LS2v/LS2v.nc"),opt)